requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0

# ML and embeddings
sentence-transformers>=2.3.1
faiss-cpu>=1.9.0
torch>=2.1.0
transformers>=4.36.0

# Gemini API
google-generativeai>=0.3.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Optional for PDF generation
markdown>=3.5.0
weasyprint>=60.0
